requirements.txt
astroid==1.6.1
confusable-homoglyphs==3.0.0
dj-database-url==0.4.2
Django==1.11
django-bootstrap3==9.1.0
django-like-button==0.0.7
django-likes==1.11
django-registration==2.4.1
django-secretballot==1.0.0
djangorestframework==3.8.2
gunicorn==19.7.1
isort==4.3.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.0.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
pylint==1.8.2
python-decouple==3.1
pytz==2017.3
six==1.11.0
whitenoise==3.3.1
wrapt==1.10.11
